Move the export logic onto methods on the ApiV2 class
making it easier to invoke programmatically.
